Create sonic_platform.py
