+++ src/sonic-config-engine/setup.py
@@ -15,7 +15,7 @@ setup(name='sonic-config-engine',
       author='Taoyu Li',
       author_email='taoyl@microsoft.com',
       url='https://github.com/Azure/sonic-buildimage',
-      py_modules=['minigraph', 'openconfig_acl'],
+      py_modules=['minigraph', 'openconfig_acl', 'sonic_platform'],
       scripts=['sonic-cfggen', 'translate_acl'],
       install_requires=['lxml', 'jinja2', 'netaddr', 'ipaddr', 'pyyaml', 'pyangbind'],
       test_suite='setup.get_test_suite',
